change to pnpm and clean-up of structure
