+++ README.md
@@ -1,0 +1,11 @@
+# Nyx-Athena Editor | NAE
+
+## name origin
+
+- Nyx: dark theme -> goddess of the night -> greek mythology
+- Athena: for the ones with wisdom -> goddess of the wisdom, intelligence -> greek mythology
+- Editor: because its lacking an integrated development enviroment -> maybe in the future?
+
+## installation instructions
+
+run `pnpm install` in top level directory
