+++ package.json
@@ -1,11 +1,7 @@
 {
-	"scripts": {
-		"tauri": "tauri"
-	},
-	"devDependencies": {
-		"@tauri-apps/cli": "^1.0.0-beta.10"
-	},
-	"dependencies": {
-		"@tauri-apps/api": "^1.0.0-beta.8"
-	}
+	"name": "athenaide-monorepo",
+	"workspaces": [
+		"src-electron/*",
+		"src-svelte/*"
+	]
 }
