+++ pnpm-workspace.yaml
@@ -1,0 +1,5 @@
+packages:
+  # all packages in subdirs of packages/ and components/
+  - "src-svelte/**"
+  - "src-electron/**"
+  # exclude packages that are inside test directories
