+++ src-electron/assets/loading_window/icon.png
