+++ src-electron/assets/loading_window/index.html
