--- src-electron/preLoadPlugin.ts
@@ -1,10 +1,0 @@
-import { pluginLoader } from "./pluginLoader";
-
-function main() {
-	console.log("in forked process");
-	console.log(process.versions);
-	// pluginLoader();
-
-}
-
-main();
