+++ src-svelte/package.json
@@ -53,7 +53,6 @@
 		"refractor": "^4.2.1",
 		"svelte-slimscroll": "^1.0.6",
 		"three": "^0.132.2",
-		"tree-sitter": "^0.20.0",
 		"tree-sitter-typescript": "^0.19.0",
 		"troika-three-text": "^0.42.0",
 		"web-tree-sitter": "^0.19.4"
